Rename to python_mip_to_ommx.py and fix imports
